Promotes mock tabular to package
