+++ tests/unit/tabular/test_to_tabular.py
@@ -1,18 +1,17 @@
-from typing import Callable
 
 import pytest
 
-from . import Config
+from gen3_util.meta.tabular import Config, to_tabular
 
 
-def test_to_tabular_default_resource(specimen: dict, to_tabular: Callable):
+def test_to_tabular_default_resource(specimen: dict):
     """Should find and use default config."""
     config = Config(default={'Specimen': ['id']})
     tabular = to_tabular(specimen, config, 'test-test')
     assert tabular == {'id': specimen['id']}
 
 
-def test_to_tabular_program_resource(specimen: dict, to_tabular: Callable):
+def test_to_tabular_program_resource(specimen: dict):
     """Should find and use program config."""
     config = Config(default={'foo': []}, programs={'test': {'Specimen': ['id']}})
     tabular = to_tabular(specimen, config, 'test-test')
@@ -23,7 +22,7 @@ def test_to_tabular_program_resource(specimen: dict, to_tabular: Callable):
         to_tabular(specimen, config, 'non_existent_program-project')
 
 
-def test_to_tabular_project_resource(specimen: dict, to_tabular: Callable):
+def test_to_tabular_project_resource(specimen: dict):
     """Should find and use project config, degrade to program, then degrade to default."""
     config = Config(default={'Specimen': ['receivedTime']}, programs={'test': {'Specimen': ['id']}}, projects={'test-test': {'Specimen': ['id', 'status']}})
 
@@ -37,7 +36,7 @@ def test_to_tabular_project_resource(specimen: dict, to_tabular: Callable):
     assert tabular == {'receivedTime': specimen['receivedTime']}
 
 
-def test_to_tabular_deep_field_list(specimen: dict, to_tabular: Callable, specimen_fields: list[str]):
+def test_to_tabular_deep_field_list(specimen: dict, specimen_fields: list[str]):
     """Should preserve deep fields."""
 
     config = Config(default={'Specimen': specimen_fields})
