+++ tests/unit/tabular/test_validate.py
@@ -1,13 +1,14 @@
-from typing import Callable
+
+from gen3_util.meta.tabular import validate
 
 
-def test_validate_simple(specimen: dict, validate: Callable):
+def test_validate_simple(specimen: dict):
     """Should validate."""
     result = validate(specimen)
     assert result.exception is None
 
 
-def test_validate_no_resource_type(specimen: dict, validate: Callable):
+def test_validate_no_resource_type(specimen: dict):
     """Should not validate."""
     del specimen['resourceType']
     result = validate(specimen)
@@ -15,7 +16,7 @@ def test_validate_no_resource_type(specimen: dict, validate: Callable):
     assert 'resourceType' in str(result.exception)
 
 
-def test_validate_no_bad_date_time(specimen: dict, validate: Callable):
+def test_validate_no_bad_date_time(specimen: dict):
     """Should not validate."""
     specimen['receivedTime'] = '1234ABCD'
     result = validate(specimen)
